back_office/requirements.txt
# Django
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0

# JWT Authentication
djangorestframework-simplejwt==5.4.0

# Database
psycopg2-binary==2.9.10
mysqlclient==2.2.6

# Environment variables
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
